% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc_functions.R
\name{date_test}
\alias{date_test}
\title{Compare the values of two date items}
\usage{
date_test(data, varm1, varj1, comparison = "smaller", varm2, varj2)
}
\arguments{
\item{data}{A neps field-data dataframe.}

\item{varm1}{A String that holds the name of a monthly date variable.}

\item{varj1}{A String that holds the name of a yearly date variable.}

\item{comparison}{A String that holds a comparison operator. Valid Operators: 'smaller', 'smaller_equal', 'larger', 'larger_equal' and 'equal'.}

\item{varm2}{A String that holds the name of a monthly date variable.}

\item{varj2}{A String that holds the name of a monthly date variable.}
}
\description{
It compares two variables (month and year) of date item 1 with two variables of date item 2 (month and year).
Valid comparison operators are: 'smaller', 'smaller_equal', 'larger', 'larger_equal' and 'equal'.
Usually used to test ra-tags, where we compare two or more date variables against each other. The function mostly has to be used more than once on order to test a range tag completly.
}
