% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc_functions.R
\name{date_test_complex}
\alias{date_test_complex}
\title{Test event or episode date variables}
\usage{
date_test_complex(
  data,
  pm1,
  pj1,
  pm2,
  pj2,
  varm1,
  varj1,
  varm2 = NULL,
  varj2 = NULL
)
}
\arguments{
\item{data}{A neps field-data dataframe.}

\item{pm1}{A String that holds the name of a monthly date variable which value will be used to test the date ranges of the target variable.}

\item{pj1}{A String that holds the name of a yearly date variable which value will be used to test the date ranges of the target variable.}

\item{pm2}{A String that holds the name of a monthly date variable which value will be used to test the date ranges of the target variable.}

\item{pj2}{A String that holds the name of a yearly date variable which value will be used to test the date ranges of the target variable.}

\item{varm1}{A String that holds the name of a monthly date variable which is usually the start date of the episode}

\item{varj1}{A String that holds the name of a yearly date variable which is usually the start date of the episode}

\item{varm2}{A String that holds the name of a monthly date variable which is usually the end date of the episode. This second date variable is only used when testing episode dates and not for event dates}

\item{varj2}{A String that holds the name of a yearly date variable which is usually the end date of the episode This second date variable is only used when testing episode dates and not for event dates}
}
\value{
either an error with an informative message about the type of error and the caseids involved or a success output print which wont stop the script
}
\description{
date range testing: supply 7 arguments when testing events, supply 9 arguments when testing episodes. After data argument, start with the two check dates, e.g. intmPRE, intjPRE, intm, intj. Then add more arguments, either an event date to test, e.g. fphm, fphj or an episode like ezstm, ezstj, ezendm, ezendj.
}
\examples{
test_data <- data.frame(caseid = c(12393535), intmPRE = c(-97), intjPRE = c(2020),
ezstm = c(-97), ezstj = c(2024), ezendm = c(6), ezendj = c(-97), intm = c(12), intj = c(2024))
date_test_complex(test_data,"intmPRE","intjPRE","intm","intj","ezstm","ezstj","ezendm","ezendj")

}
