% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc_functions.R
\name{expand}
\alias{expand}
\title{Expand episode data to monthly structure}
\usage{
expand(data, duration)
}
\arguments{
\item{data}{A dataframe}

\item{duration}{A String with a variable name that will be used for expanding. Usually a numerical duration variable.}
}
\description{
Expand episode data to monthly structure
}
