% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nepstest-package.R
\docType{package}
\name{nepstest-package}
\alias{nepstest}
\alias{nepstest-package}
\title{nepstest: What the Package Does (One Line, Title Case)}
\description{
Functions for testing NEPS field-data, in particular af-tag, ac-tag, ra-tag.
}
\author{
\strong{Maintainer}: Alexander Helbig \email{alexander.helbig@wzb.eu} (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID})

}
\keyword{internal}
