% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc_functions.R
\name{replace_values_with_na}
\alias{replace_values_with_na}
\title{Set specific values to NA
If only the data argument is supplied, the function will set specific values in all variables to NA.}
\usage{
replace_values_with_na(
  data,
  vars = NULL,
  values_to_replace = c(seq(-99, -90), seq(-56, -51), seq(-29, -22))
)
}
\arguments{
\item{data}{A neps field-data dataframe.}

\item{vars}{A string vector with variable names where specific values will be replaced with NA.}

\item{values_to_replace}{A numerical vector with values that will be replaced with NA. By default most of the standard NEPS missings will be converted to NA.}
}
\description{
Set specific values to NA
If only the data argument is supplied, the function will set specific values in all variables to NA.
}
