% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ac_tags.R
\name{ac_test}
\alias{ac_test}
\title{Tests if target variable equals target value}
\usage{
ac_test(
  data,
  target_var,
  target_value,
  condition_var = NULL,
  operator = "equal",
  condition_value = NULL
)
}
\arguments{
\item{data}{A neps field-data dataframe.}

\item{target_var}{A string that represents a variable within the dataframe.}

\item{target_value}{A string that can either be a digit or a variable name present in the dataframe.}

\item{condition_var}{A string representing a variable name to be used as a conditional variable.}

\item{operator}{A string that specifies the operator to be used for the condition. Acceptable operators include: "equal", "unequal", "greater", "smaller", "greaterequal", "smallerequal", and "inlist".}

\item{condition_value}{A string that contains a digit to be used as the conditional value.}
}
\description{
Tests if target variable equals target value
}
