% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/af_tags.R
\name{af_test_complex}
\alias{af_test_complex}
\title{Tests complex af-tag filters}
\usage{
af_test_complex(
  data,
  test_var,
  target_var,
  overfiltered_vars,
  condition_var = NULL,
  operator = "equal",
  condition_value = NULL,
  empty_as_na = TRUE
)
}
\arguments{
\item{data}{A neps field-data dataframe.}

\item{test_var}{A string that represents the variable where we want to test the af-tag.}

\item{target_var}{A string that represents the variable where the af-tag is pointing too.}

\item{overfiltered_vars}{A string containing a single variable or a character vector of variables that will be skipped according to the af-tag and must be NA.}

\item{condition_var}{A string representing a variable name to be used as a conditional variable.}

\item{operator}{A string that specifies the operator to be used for the condition. Acceptable operators include: "equal", "unequal", "greater", "smaller", "greaterequal", "smallerequal", and "inlist".}

\item{condition_value}{A string that contains a digit to be used as the conditional value.}

\item{empty_as_na}{Logical, if TRUE empty strings ("") in character variables will be treated as NA. Default is TRUE}
}
\description{
This function is designed for testing af-tags that filter from one item to another item by skipping the items in between (under condition x).
It tests for NA equality between test_var and target_var and additionally if all vars in overfiltered_vars are NA.
}
