% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/af_tags.R
\name{af_test_simple}
\alias{af_test_simple}
\title{Tests for NA equality between var1 and var2}
\usage{
af_test_simple(
  data,
  var1,
  var2,
  condition_var = NULL,
  operator = "equal",
  condition_value = NULL
)
}
\arguments{
\item{data}{A neps field-data dataframe.}

\item{var1}{A string that represents the variable where we want to test the af-tag.}

\item{var2}{A string that represents the variable where the af-tag is pointing too. Usually the following item.}

\item{condition_var}{A string representing a variable name to be used as a conditional variable.}

\item{operator}{A string that specifies the operator to be used for the condition. Acceptable operators include: "equal", "unequal", "greater", "smaller", "greaterequal", "smallerequal", and "inlist".}

\item{condition_value}{A string that contains a digit to be used as the conditional value.}
}
\description{
This function is designed for testing af-tags that filter from one item to the following item (under condition x).
}
