% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/na_checks.R
\name{no_na_in_vars}
\alias{no_na_in_vars}
\title{Checks if there are no NA values in specified var_names}
\usage{
no_na_in_vars(
  data,
  var_names,
  condition_var = NULL,
  operator = "equal",
  condition_value = NULL
)
}
\arguments{
\item{data}{A neps field-data dataframe.}

\item{var_names}{A character vector with variables that will be checked to ensure there are no NA values.}

\item{condition_var}{A string representing a variable name to be used as a conditional variable.}

\item{operator}{A string that specifies the operator to be used for the condition. Acceptable operators include: "equal", "unequal", "greater", "smaller", "greaterequal", "smallerequal", and "inlist".}

\item{condition_value}{A string that contains a digit to be used as the conditional value.}
}
\description{
Checks if there are no NA values in specified var_names
}
