% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc_functions.R
\name{replace_season_codes}
\alias{replace_season_codes}
\title{Replace season codes in date variables with corresponding months}
\usage{
replace_season_codes(
  data,
  vars = NULL,
  values_to_replace = c(21, 24, 27, 30, 32)
)
}
\arguments{
\item{data}{A neps field-data dataframe.}

\item{vars}{A string vector with variable names where season codes will be replaced with actual months.}

\item{values_to_replace}{Values that will be replaced with months. Only need to be edited in case of new season codes.}
}
\description{
This function will transform neps season codes to month values by subtracting 20 from each season code.
E.g. code 24 - "Spring" will be 4 - "April".
}
