% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc_functions.R
\name{var_exists}
\alias{var_exists}
\title{Test if variables are existent in the data}
\usage{
var_exists(data, vars)
}
\arguments{
\item{data}{A neps field-data dataframe.}

\item{vars}{A string vector with variable names. It will be checked if these exist in the dataframe.}
}
\description{
This function can be used to test if a bunch of Hilfsvariable are existent in dataframe where we usually dont have any filtering in the instruction sheets.
}
